% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBOE_VIX.R
\name{CBOE_delta_K}
\alias{CBOE_delta_K}
\title{Calculate the average strike price distance variable \mjseqn{\Delta K_i} from the CBOE VIX calculation.}
\usage{
CBOE_delta_K(K)
}
\arguments{
\item{K}{A \code{numeric vector} of strike prices}
}
\value{
Returns a \code{numeric vector} giving the \mjseqn{\Delta K_i} variable of the CBOE VIX calculation.
}
\description{
Following the \href{https://www.cboe.com/micro/vix/vixwhite.pdf}{VIX whitepaper} this function
calculates \mjseqn{\Delta K_i} as \emph{"half the difference between the strike prices on either side of
\mjseqn{K_i}"}:
\loadmathjax
\mjsdeqn{\Delta K_i := \begin{cases} K_1 - K_0 \qquad \; \; if \; i = 0 \\\\\
K_N - K_{N-1} \; \; \; if \; i = N \\\\\
\frac{K_{i+1} - K_{i-1}}{2} \qquad \; else
\end{cases}}
}
\examples{

strikes <- c(10, 12.5, c(1:10)*5 + 10, 62.5, 65)
CBOE_delta_K(K = strikes)
}
