% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBOE_VIX.R
\name{CBOE_F_0}
\alias{CBOE_F_0}
\title{Calculate the theoretical at-the-money forward \mjseqn{F_0} from the CBOE VIX calculation.
\loadmathjax}
\usage{
CBOE_F_0(option_quotes, R, maturity)
}
\arguments{
\item{option_quotes}{A \code{data.table} or "nest" of option quotes with three columns:
\itemize{
\item {\strong{K} (\code{numeric})}{ - strike price in ascending order}
\item {\strong{c} (\code{numeric})}{ - call option price}
\item {\strong{p} (\code{numeric})}{ - put option price}
}}

\item{R}{\code{numeric scalar} giving the risk-free rate(s) \mjseqn{R}}

\item{maturity}{\code{numeric scalar} giving the time(s) to maturity \mjseqn{T}}
}
\value{
Returns a \code{numeric scalar}, giving the theoretical at-the-money
forward \mjseqn{F_0}
}
\description{
Following the \href{https://www.cboe.com/micro/vix/vixwhite.pdf}{VIX whitepaper} this function
calculates \mjseqn{F_0} as:

\mjsdeqn{F_0 := Strike Price + e^{RT} (Call Price - Put Price)}

\mjseqn{R}is the risk-free-rate (in decimal) for the corresponding time-to-maturity
\mjseqn{T} (in years). The \mjseqn{Call Price} and \mjseqn{Put Price} are those where their absolute
difference is smallest.
}
\examples{

library(R.MFIV)

nest <- option_dataset$option_quotes[[1]]

CBOE_F_0(option_quotes = nest,
         R = 0.005,
         maturity = 0.07)
}
