% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBOE_VIX.R
\name{CBOE_interpolation_terms}
\alias{CBOE_interpolation_terms}
\title{Determine the expiration "term" used in the linear interpolation of the CBOE VIX}
\usage{
CBOE_interpolation_terms(maturity, date_t, date_exp, method)
}
\arguments{
\item{maturity}{\code{numeric scalar} giving the time to maturity \mjseqn{T} in years}

\item{date_t}{(For the monthly method) \code{date scalar} giving the date of the option quotation}

\item{date_exp}{(For the monthly method) \code{date scalar} giving the date of the option expiration}

\item{method}{A \code{string scalar}, either \code{"weekly"} or \code{"monthly"} for the
respective method.}
}
\value{
Returns a \code{numeric scalar}: \code{1} for the "near-term" and \code{2} for the "next-term".
If the maturity doesn't fall inside either category, \code{NA} is returned.
}
\description{
Provide either \code{maturity} or \code{date_t} and \code{date_exp}

This function determines the interpolation terms used for the VIX.
It provides terms for the two different interpolation techniques used by the CBOE:
\itemize{
  \item {\strong{2003 VIX} (monthly):} using monthly options (see the \href{https://web.archive.org/web/20091231021416/https://cdn.cboe.com/resources/vix/vixwhite.pdf}{2009 CBOE Whitepaper})
  \item {\strong{2014 VIX} (weekly):} using weekly  options (see the \href{https://cdn.cboe.com/resources/vix/vixwhite.pdf}{2019 VIX whitepaper})
}

Both methods rely on a "near-term" and a "next-term" contract.
}
\examples{

library(R.MFIV)

## Weekly method
CBOE_interpolation_terms(25/365, method = "weekly")

## Monthly method
t <- lubridate::ymd("2020-01-02")
exp <- lubridate::ymd("2020-02-21")
CBOE_interpolation_terms(date_t = t, date_exp = exp, method = "monthly")

}
