% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBOE_VIX.R
\name{CBOE_K_0}
\alias{CBOE_K_0}
\title{Calculate the theoretical at-the-money forward \mjseqn{K_0} from the CBOE VIX
calculation.
\loadmathjax}
\usage{
CBOE_K_0(option_quotes, F_0)
}
\arguments{
\item{option_quotes}{A \code{data.table} or "nest" of option quotes with three columns:
\itemize{
  \item {\strong{K} (\code{numeric})} - strike price in ascending order
  \item {\strong{c} (\code{numeric})} - call option price
  \item {\strong{p} (\code{numeric})} - put option price
}}

\item{F_0}{\code{numeric scalar}, giving the  theoretical at-the-money forward \mjseqn{F_0}
(see \code{\link{CBOE_F_0}})}
}
\value{
Returns a \code{numeric scalar}, giving the theoretical at-the-money
 strike \mjseqn{K_0}
}
\description{
Following the \href{https://cdn.cboe.com/resources/vix/vixwhite.pdf}{VIX whitepaper}
\mjseqn{K_0} is defined as the strike price directly smaller or equal to the
theoretical at-the-money forward price \mjseqn{F_0}
}
\examples{

library(R.MFIV)

nest <- option_dataset$option_quotes[[1]]

F_0 <- CBOE_F_0(option_quotes = nest,
         R = 0.005,
         maturity = 0.07)

K_0 <- CBOE_K_0(option_quotes = nest,
         F_0 = F_0)
}
