% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBOE_VIX.R
\name{CBOE_sigma_sq}
\alias{CBOE_sigma_sq}
\title{Calculate the CBOE VIX model free variance \mjseqn{\sigma^2}}
\usage{
CBOE_sigma_sq(sel_option_quotes, K_0, F_0, maturity, R)
}
\arguments{
\item{sel_option_quotes}{A \code{data.table} or "nest" of option quotes as selected by
\code{\link{CBOE_option_selection}}}

\item{K_0}{\code{numeric scalar}, giving the theoretical at-the-money strike price
(see \code{\link{CBOE_K_0}})}

\item{F_0}{\code{numeric scalar}, giving the  theoretical at-the-money forward \mjseqn{F_0}
(see \code{\link{CBOE_F_0}})}

\item{maturity}{\code{numeric scalar} giving the time to maturity \mjseqn{T} in years}

\item{R}{\code{numeric scalar} giving the risk-free rate \mjseqn{R} corresponding to the
maturity \mjseqn{T} in decimal}
}
\value{
Returns a \code{numeric scalar}: the model-free implied volatility \mjseqn{\sigma^2} as
per the CBOE formula above.
}
\description{
This function performs the CBOE VIX model-free implied variance calculation
according to the following formula from the \href{https://www.cboe.com/micro/vix/vixwhite.pdf}{2019 VIX whitepaper}:
\loadmathjax
\mjsdeqn{\sigma^2 = \frac{2}{T} \left(\sum_i \frac{\Delta K_i}{K_i^2} Q(K_i) e^{rT} \right) - \frac{1}{T} \left( \frac{F_0}{K_0} - 1 \right)^2}
It uses \code{\link{CBOE_delta_K}} internally to derive the weights \mjseqn{\Delta K_i}
}
\examples{

library(R.MFIV)
nest <- CBOE_option_selection(option_dataset$option_quotes[[1]],
                              147)
CBOE_sigma_sq(sel_option_quotes = nest,
              K_0 = 147,
              F_0 = 147.5,
              maturity = 0.07,
              R = 0.005)
}
