% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBOE_VIX.R
\name{CBOE_VIX_index}
\alias{CBOE_VIX_index}
\title{Calculates the linear VIX interpolation and returns the
annualised VIX index in percentage points}
\usage{
CBOE_VIX_index(maturity, sigma_sq)
}
\arguments{
\item{maturity}{\code{numeric vector of length two} giving the time to maturity of the
"near-term" and "next-term" options in years (\mjseqn{T_1} and \mjseqn{T_2} respectively)}

\item{sigma_sq}{\code{numeric vector of length two} giving the CBOE model-free implied volatility
of the "near-term" and "next-term" options(\mjseqn{\sigma^2_1} and \mjseqn{\sigma^2_2} respectively).
Also see \code{\link{CBOE_sigma_sq}}}
}
\value{
Returns a \code{numeric scalar} giving the VIX index value
}
\description{
\loadmathjax
Following the \href{https://cdn.cboe.com/resources/vix/vixwhite.pdf}{VIX whitepaper}, the VIX index is
calculated as a linear interpolation. This function uses the following formula:
\mjsdeqn{\sigma_{VIX} =  100 \sqrt{ \big( \omega T_1 \sigma_1^2 + (1- \omega) T_2 \sigma_2^2 \big) \frac{525,600}{43,200} }}
where the subscripts \mjseqn{1} and \mjseqn{2} indicate the near-and next-term options, \mjseqn{T_\cdot}
to the time to expiration in years and the \mjseqn{\omega} to the linear interpolation weights.
}
\examples{

library(R.MFIV)

CBOE_VIX_index(maturity = c(0.074, 0.09),
               sigma_sq = c(0.3, 0.5))


}
