% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBOE_VIX.R
\name{CBOE_VIX_vars}
\alias{CBOE_VIX_vars}
\title{Calculate all variables needed for the calculation of the CBOE VIX}
\usage{
CBOE_VIX_vars(option_quotes, R, maturity, ret_vars = F)
}
\arguments{
\item{option_quotes}{A \code{data.table} or "nest" of option quotes with three columns:
\itemize{
\item {\strong{K} (\code{numeric})}{ - strike price in ascending order}
\item {\strong{c} (\code{numeric})}{ - call option price}
\item {\strong{p} (\code{numeric})}{ - put option price}
}}

\item{R}{\code{numeric scalar} giving the risk-free rate \mjseqn{R} corresponding to the
maturity \mjseqn{T} in decimal}

\item{maturity}{\code{numeric scalar} giving the time to maturity \mjseqn{T} in years}

\item{ret_vars}{A \code{logical scalar} - if true, all VIX variables are returned, else only \mjseqn{\sigma^2}
is returned.}
}
\value{
Returns either a \code{numeric scalar} giving \mjseqn{\sigma^2} or a \code{list} with all variables
involved in the calculation:
\itemize{
\item {\strong{F_0} (\code{numeric})}{ - theoretical at-the money forward \mjseqn{F_0}
(see \code{\link{CBOE_F_0}})}
\item {\strong{K_0} (\code{numeric})}{ - theoretical at-the money strike \mjseqn{K_0}
(see \code{\link{CBOE_K_0}})}
\item {\strong{n_put_raw} (\code{numeric})}{ - number of put options before option selection
(see \code{\link{CBOE_option_selection}})}
\item {\strong{n_call_raw} (\code{numeric})}{ - number of call options before option selection
(see \code{\link{CBOE_option_selection}})}
\item {\strong{n_put} (\code{numeric})}{ - number of put options after option selection}
\item {\strong{n_call} (\code{numeric})}{ - number of call options after option selection}
\item {\strong{sigma_sq} (\code{numeric})}{ - squared model free implied volatility \mjseqn{\sigma^2}
(see \code{\link{CBOE_sigma_sq}})}
}
}
\description{
This is a wrapper around the \code{CBOE_...} functions that performs the calculation
of all variables required for the calculation of the squared model free implied volatility (\mjseqn{\sigma^2})
as per the \href{https://www.cboe.com/micro/vix/vixwhite.pdf}{VIX whitepaper}:

\loadmathjax
\mjsdeqn{\sigma^2 = \frac{2}{T} \left(\sum_i \frac{\Delta K_i}{K_i^2} Q(K_i) e^{rT} \right) - \frac{1}{T} \left( \frac{F_0}{K_0} - 1 \right)^2}
}
\examples{

library(R.MFIV)

nest <- option_dataset$option_quotes[[1]]
CBOE_VIX_vars(option_quotes = nest,
              R = 0.005,
              maturity = 0.07,
              ret_vars = TRUE)

}
