% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cmt_dataset}
\alias{cmt_dataset}
\title{CMT rate date obtained from the US Treasury as of \code{lubridate::now()}}
\format{
A \code{data.table]
}
\source{
\href{https://www.treasury.gov/resource-center/data-chart-center/interest-rates/pages/TextView.aspx?data=yieldAll}{US Treasury website}
}
\usage{
cmt_dataset
}
\description{
CMT rate date obtained from the US Treasury as of \code{lubridate::now()}
}
\keyword{datasets}
