% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy}
\alias{dummy}
\title{Dummy function
\loadmathjax
Performs \mjseqn{a + b}}
\usage{
dummy(a, b)
}
\arguments{
\item{a}{numeric scalar}

\item{b}{numeric scalar}
}
\value{
a numeric
}
\description{
Dummy function
\loadmathjax
Performs \mjseqn{a + b}
}
\examples{

dummy(1, 1)
}
