% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-free-rate.R
\name{interpolate_rfr}
\alias{interpolate_rfr}
\title{Calculate risk-free-rates through cubic-spline interpolation using constant maturity
treasury (CMT) rates data.}
\usage{
interpolate_rfr(cmt_data = NULL, date, exp, ret_table = FALSE)
}
\arguments{
\item{cmt_data}{A \code{data.table} containing CMT quotes. See \code{\link{cmt_dataset}} for an example
and \code{\link{scrape_cmt_data}} for a way to obtain it.}

\item{date}{A \code{date vector} of starting dates.}

\item{exp}{A \code{date vector} of expiration dates.}

\item{ret_table}{\code{Logical scalar}, indicating whether to return only the risk free rate(s) or a
table with the original \code{date}, \code{exp} and the corresponding risk-free-rates}
}
\value{
Returns either a \code{numeric vector} of risk free rates or a \code{data.table} containing:
\itemize{
  \item{\strong{date}(\code{date})} - the provided starting dates
  \item{\strong{exp} (\code{date})} - the provided expiration dates
  \item{\strong{R}   (\code{numeric})} - the corresponding risk-free interest rates in \emph{decimal}
}
}
\description{
According to the \href{https://cdn.cboe.com/resources/vix/vixwhite.pdf}{2019 VIX whitepaper},
\emph{"the risk-free interest rates, R1 and R2, are yields based on U.S. Treasury yield curve
rates (commonly referred to as "Constant Maturity Treasury", rates or CMTs), to which a
cubic spline is applied to derive yields on the expiration dates of relevant SPX options.
As such, the VIX Index calculation may use different risk-free interest rates for near- and
next-term options. Note in this example, T2 uses a value of 900 for Settlement day, which
reflects the 4:00 p.m. ET expiration time of the next-term SPX Weeklys options."}
}
\examples{

## Using internal package data
library(lubridate)

## Single values
interpolate_rfr(date = ymd("2020-01-02"), exp = date("2020-03-02"))


## Multiple values
dates <- ymd("2020-01-06") + days(1:4)
exps <-  ymd("2020-03-02") + days(1:4)

interpolate_rfr(date = dates, exp = exps)

## With output table
interpolate_rfr(date = dates, exp = exps, ret_table = TRUE)

}
