% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JandT_smoothing.R
\name{JandT_2007_sigma_sq}
\alias{JandT_2007_sigma_sq}
\title{Calculate the MFIV according to the Jiang & Tian (2007) paper}
\usage{
JandT_2007_sigma_sq(smooth_option_quotes, K_0, maturity, R)
}
\arguments{
\item{smooth_option_quotes}{A \code{data.table} or "nest" of option quotes as returned
from \code{\link{JandT_2007_smoothing_method}}}

\item{K_0}{\code{numeric scalar}, giving the theoretical at-the-money strike price
(see \code{\link{CBOE_K_0}})}

\item{maturity}{\code{numeric scalar} giving the time to maturity \mjseqn{T} in years}

\item{R}{\code{numeric scalar} giving the risk-free rate \mjseqn{R} corresponding to the
maturity \mjseqn{T} in decimal}
}
\value{
Returns a \code{numeric scalar}: the model-free implied volatility \mjseqn{\sigma^2} as
per the CBOE formula above.
}
\description{
This function performs the model-free implied variance calculation
according to the following formula from Jiang & Tian (2007):
\loadmathjax
\mjsdeqn{V = \frac{2}{T} \exp(R T) (A + B)  }
\mjsdeqn{A = \sum_{i \leq 0} \frac{\Delta \hat{K}_i}{2} \left( \frac{P^{EX} (\hat{K}_i, T)}{\hat{K}^2_i} + \frac{P^{EX} (\hat{K}_j, T)}{\hat{K}^2_j}  \right)  }

\mjsdeqn{B = \sum_{i \leq 0} \frac{\Delta \hat{K}_i}{2} \left( \frac{C^{EX} (\hat{K}_i, T)}{\hat{K}^2_i} + \frac{C^{EX} (\hat{K}_j, T)}{\hat{K}^2_j}  \right)  }
\mjsdeqn{ \Delta \hat{K}_i = \hat{K}_i - \hat{K}_j }
\mjsdeqn{j = i-1}
}
\examples{

library(R.MFIV)

## LOAD EXAMPLE OPTION_QUOTES
nest <- option_dataset$option_quotes[[1]]

## EXTRAPOLATE DATA
smooth_nest <- JandT_2007_smoothing_method(option_quotes = nest,
                                           maturity = 0.008953152,
                                           K_0 = 147,
                                           price = 147.39,
                                           R = 0.008325593,
                                           F_0 = 147.405)
## CALCULATE MFIV
sigma_sq <- JandT_2007_sigma_sq(smooth_option_quotes = smooth_nest,
                                K_0 = 147,
                                maturity = 0.008953152,
                                R = 0.008325593)
}
\references{
\doi{10.3905/jod.2007.681813} Jiang & Tian (2007) - Extracting Model-Free
Volatility
}
