% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JandT_smoothing.R
\name{JandT_2007_smoothing_method}
\alias{JandT_2007_smoothing_method}
\title{Extend and fill in the option quotes using Jiang & Tian (2005, 2007)
Implied Volatility Curve Fitting
\loadmathjax}
\usage{
JandT_2007_smoothing_method(option_quotes, K_0, price, R, maturity, F_0)
}
\arguments{
\item{option_quotes}{A \code{data.table} or "nest" of option quotes with three columns:
\itemize{
\item {\strong{K} (\code{numeric})}{ - strike price in ascending order}
\item {\strong{c} (\code{numeric})}{ - call option price}
\item {\strong{p} (\code{numeric})}{ - put option price}
}}

\item{K_0}{\code{numeric scalar}, giving the theoretical at-the-money strike price
(see \code{\link{CBOE_K_0}})}

\item{price}{\code{numeric scalar} giving the underlying stock price.}

\item{R}{\code{numeric scalar} giving the risk-free rate \mjseqn{R} corresponding to the
maturity \mjseqn{T} in decimal}

\item{maturity}{\code{numeric scalar} giving the time to maturity \mjseqn{T} in years}

\item{F_0}{\code{numeric scalar}, giving the  theoretical at-the-money forward \mjseqn{F_0}
(see \code{\link{CBOE_F_0}})}
}
\value{
Returns a \code{data.table} of intra- and extrapolated option quotes with the
following columns:
\itemize{
\item {\strong{K} (\code{numeric})}{ - strike price in ascending order}
\item {\strong{Q} (\code{numeric})}{ - OTM option}
}
}
\description{
Extend and fill in the option quotes using Jiang & Tian (2005, 2007)
Implied Volatility Curve Fitting
\loadmathjax
}
\details{
The CBOE VIX calculation method relies on a wide option quotes over the range of
strike prices. However, the actual data might not yield enough option quotes. If
the quotes are spaced apart too far and have a limited range over the strike prices,
this introduces errors (see Jiang & Tian (2005, 2007)).

Jiang & Tian (2005, 2007) also propose a smoothing method: option quotes are
transformed in to Black \& Scholes implied volatilities, intra- and extrapolated,
an then transformed back into option prices. This methodology is provided by this
function.

After application of this method, the returned option_quotes are spaced apart by
\mjseqn{0.35} "SD Units", which translates into option price increments of
\mjseqn{SD \cdot \sqrt{maturity} \cdot price \cdot 0.35}, where \mjseqn{SD}
refers to a B&S implied volatility of the at-the-money option. The range of the
returned data spans \mjseqn{\pm 15} SD units.
}
\examples{

library(R.MFIV)

## LOAD EXAMPLE OPTION_QUOTES
nest <- option_dataset$option_quotes[[1]]

## EXTRAPOLATE DATA
JandT_2007_smoothing_method(option_quotes = nest,
                            maturity = 0.008953152,
                            K_0 = 147,
                            price = 147.39,
                            R = 0.008325593,
                            F_0 = 147.405)

}
\references{
\href{https://doi.org/10.1093/RFS\%2FHHI027}{Jiang & Tian (2005) -
The Model-Free Implied Volatility and Its Information Content}

\href{https://doi.org/10.3905/jod.2007.681813}{Jiang & Tian (2007) -
Extracting Model-Free Volatility
from Option Prices: An Examination of the VIX Index}
}
