% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{option_dataset}
\alias{option_dataset}
\title{Example option quote dataset.}
\format{
A \code{data.table} with the following columns:
\itemize{
\item{\strong{ticker}   (\code{character})}{ - the underlying stock ticker symbol}
\item{\strong{t}  (\code{datetime})}{ - the time of quotation}
\item{\strong{exp}  (\code{date})}{ - the expiration date}
\item{\strong{price}  (\code{numeric})}{ - the underlying stock price}
\item{\strong{option_quotes}  (\code{data.table})}{ - a "nest" of option quotes.
Each cell in this column contains a \code{data.table} with the following columns:
\itemize{
\item{\strong{K}  (\code{numeric})}{ - the strike price}
\item{\strong{c}  (\code{numeric})}{ - the corresponding call price}
\item{\strong{p}  (\code{numeric})}{ - the correspondign put price}
}
}
}
}
\usage{
option_dataset
}
\description{
Example option quote dataset.
}
\keyword{datasets}
