% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{option_descriptives}
\alias{option_descriptives}
\title{Calculate descriptive variables fo a nest of option quotes}
\usage{
option_descriptives(option_quotes, K_0, R, price, maturity)
}
\arguments{
\item{option_quotes}{A \code{data.table} or "nest" of option quotes with three columns:
\itemize{
\item {\strong{K} (\code{numeric})}{ - strike price in ascending order}
\item {\strong{c} (\code{numeric})}{ - call option price}
\item {\strong{p} (\code{numeric})}{ - put option price}
}}

\item{K_0}{\code{numeric scalar}, giving the theoretical at-the-money strike price
(see \code{\link{CBOE_K_0}})}

\item{R}{\code{numeric scalar} giving the risk-free rate \mjseqn{R} corresponding to the
maturity \mjseqn{T} in decimal}

\item{price}{\code{numeric scalar} giving the underlying stock price.}

\item{maturity}{\code{numeric scalar} giving the time to maturity \mjseqn{T} in years}
}
\value{
Returns a \code{list} with the following variables:
\itemize{
\item {\strong{SD} (\code{numeric})}{- the Black Merton Scholes implied volatility for the at-the money call
(using \code{GBSVolatility} from \code{fOptions})}
\item {\strong{max_K} (\code{numeric})}{ - highest strike price for out-of-the-money calls}
\item {\strong{min_K} (\code{numeric})}{ - lowest strike price for out-of-the-money puts}
\item {\strong{mean_delta_K} (\code{numeric})}{ - average distance between strike prices}
\item {\strong{n_put} (\code{numeric})}{ - number of out-of-the money puts}
\item {\strong{n_call} (\code{numeric})}{ - number of out-of-the money calls}
}
}
\description{
Calculate descriptive variables fo a nest of option quotes
}
\examples{

library(R.MFIV)

nest <- option_dataset$option_quotes[[1]]

option_descriptives(option_quotes = nest,
                   K_0 = 147,
                   R = 0.005,
                   price = 147,
                   maturity = 0.07)

}
