% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{R.MFIV}
\alias{R.MFIV}
\title{R.MFIV: A package for handling CBOE option-data and calculation the CBOE VIX}
\description{
See the following vignettes for an explanation of how this package works:
}
