% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-free-rate.R
\name{scrape_cmt_data}
\alias{scrape_cmt_data}
\title{Scrape the CMT data form the US Treasury website}
\usage{
scrape_cmt_data(url = NULL)
}
\arguments{
\item{url}{\code{character scalar} an URL to the US Treasury website. Defaults to the
\href{https://www.treasury.gov/resource-center/data-chart-center/interest-rates/pages/TextView.aspx?data=yieldAll}{complete dataset}.
Also accepts links from this website with other selections than "All", see e.g. the example.}
}
\value{
Returns a \code{data.table} containing the following columns:
\itemize{
\item{Date (\code{date})}{ - the day of the observation}
\item{maturity A (\code{numeric}, \emph{decimal})}{ - the CMT rate for the respective maturity in decimal}
\item{...}
}
}
\description{
According to the \href{https://www.cboe.com/micro/vix/vixwhite.pdf}{2019 VIX whitepaper},
\emph{"the risk-free interest rates, R1 and R2, are yields based on U.S. Treasury yield curve
rates (commonly referred to as "Constant Maturity Treasury", rates or CMTs), to which a
cubic spline is applied to derive yields on the expiration dates of relevant SPX options.
As such, the VIX Index calculation may use different risk-free interest rates for near- and
next-term options. Note in this example, T2 uses a value of 900 for Settlement day, which
reflects the 4:00 p.m. ET expiration time of the next-term SPX Weeklys options."}

This data can be retrieved from the \href{https://www.treasury.gov/resource-center/data-chart-center/interest-rates/pages/TextView.aspx?data=yieldAll}{US Treasury website}
(link works as of 2020-10-14)
}
\examples{

## Scrape the CMT data for the current month:
scrape_cmt_data()

}
