% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{third_fridays}
\alias{third_fridays}
\title{List third fridays in the months of the period given by \code{start} and \code{end}}
\usage{
third_fridays(start, end)
}
\arguments{
\item{start}{\code{Date scalar}, giving the start date}

\item{end}{\code{Date scalar}, giving the end date}
}
\value{
Returns a \code{date vector} with all third fridays of the respective period.
}
\description{
The 2003 VIX (see the \href{https://web.archive.org/web/20091231021416/https://www.cboe.com/micro/vix/vixwhite.pdf}{2009 CBOE Whitepaper})
is based on monthly option contracts, which expire on every third friday of a month.
This function identifies monthly third fridays in order to select the correct monthly options.
}
\examples{

start <- lubridate::ymd("2020-01-01")
end <- lubridate::ymd("2020-06-01")

third_fridays(start, end)
}
