% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option-price_extrapolation.R
\name{JandT_2007_smoothing_method}
\alias{JandT_2007_smoothing_method}
\title{Extend and fill in the option quotes using Jiang & Tian (2005) Implied Volatility Curve Fitting
\loadmathjax}
\usage{
JandT_2007_smoothing_method(option_quotes, K_0, price, R, maturity, F_0)
}
\arguments{
\item{option_quotes}{A \code{data.table} or "nest" of option quotes with three columns:
\itemize{
\item {\strong{K} (\code{numeric})}{ - strike price in ascending order}
\item {\strong{c} (\code{numeric})}{ - call option price}
\item {\strong{p} (\code{numeric})}{ - put option price}
}}

\item{K_0}{\code{numeric scalar}, giving the theoretical at-the-money strike price
(see \code{\link{CBOE_K_0}})}

\item{price}{\code{numeric scalar} giving the underlying stock price.}

\item{R}{\code{numeric scalar} giving the risk-free rate \mjseqn{R} corresponding to the
maturity \mjseqn{T} in decimal}

\item{maturity}{\code{numeric scalar} giving the time to maturity \mjseqn{T} in years}

\item{F_0}{\code{numeric scalar}, giving the  theoretical at-the-money forward \mjseqn{F_0}
(see \code{\link{CBOE_F_0}})}
}
\value{
Returns a \code{data.table} of intra- and extrapolated option quotes with the
following columns:
\itemize{
\item {\strong{K} (\code{numeric})}{ - strike price in ascending order}
\item {\strong{Q} (\code{numeric})}{ - OTM option}
}
}
\description{
\mjseqn{test}
}
\examples{

library(R.MFIV)

## LOAD EXAMPLE OPTION_QUOTES
nest <- option_dataset$option_quotes[[1]]

## EXTRAPOLATE DATA
extra_data <- JandT_2007_smoothing_method(option_quotes = nest,
                                          maturity = 0.008953152,
                                          K_0 = 147,
                                          price = 147.39,
                                          R = 0.008325593,
                                          F_0 = 147.405)

}
