% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JandT_smoothing.R
\name{JandT_2007_smoothing_method}
\alias{JandT_2007_smoothing_method}
\title{Extend and fill in the option quotes using Jiang & Tian (2005, 2007)
Implied Volatility Curve Fitting
\loadmathjax}
\usage{
JandT_2007_smoothing_method(
  option_quotes,
  K_0,
  price,
  R,
  maturity,
  F_0,
  tail_length = 15,
  flat_tails = TRUE,
  increment = "min"
)
}
\arguments{
\item{option_quotes}{A \code{data.table} or "nest" of option quotes with three columns:
\itemize{
  \item {\strong{K} (\code{numeric})} - strike price in ascending order
  \item {\strong{c} (\code{numeric})} - call option price
  \item {\strong{p} (\code{numeric})} - put option price
}}

\item{K_0}{\code{numeric scalar}, giving the theoretical at-the-money strike price
(see \code{\link{CBOE_K_0}})}

\item{price}{\code{numeric scalar} giving the underlying stock price.}

\item{R}{\code{numeric scalar} giving the risk-free rate \mjseqn{R} corresponding to the
maturity \mjseqn{T} in decimal}

\item{maturity}{\code{numeric scalar} giving the time to maturity \mjseqn{T} in years}

\item{F_0}{\code{numeric scalar}, giving the  theoretical at-the-money forward \mjseqn{F_0}
(see \code{\link{CBOE_F_0}})}

\item{tail_length}{\code{numeric scalar} giving the strike-price range of the
returned option quotes in \mjseqn{SD} units (see details). If the tail length
doesn't exceed the the range of the given option quotes, no extrapolation takes place.}

\item{flat_tails}{\code{logical scalar} determining whether the extrapolation in the
tails is sloped or flat (see details)}

\item{increment}{\code{character or numeric scalar} giving the strike-price increment
of the returned option quotes in \mjseqn{SD} units (see details). Options are:
\itemize{
  \item {\strong{x} (\code{numeric})} - any number giving the strike-price increment
  in price units
  \item {\strong{"JT"} (\code{character})} - the increment from Jiang & Tian (2007)
  i.e. \mjseqn{SD \cdot \sqrt{maturity} \cdot price \cdot 0.35}
  \item {\strong{"real"} (\code{character})} - the smallest increment found in the
  "real" data
  \item {\strong{"min"} (\code{character})} - the smaller one of "JT" and "min"
}
The increment is at least 0.5, unless a numeric value is provided.}
}
\value{
Returns a \code{data.table} of intra- and extrapolated option quotes with the
following columns:
\itemize{
  \item {\strong{K} (\code{numeric})} - evenly spaced strike price in ascending order
  \item {\strong{Q} (\code{numeric})} - OTM option price
}
}
\description{
Intrapolates and extrapolates the option_quotes over the strike price range using
the Jiang & Tian (2007) smoothing method. The given quotes are transformed into
Black & Scholes implied volatilities, on which intra/extrapolation takes place.
}
\details{
Inside the given range of strike prices, (natural) cubic spline intrapolation takes
place.

The tails are extrapolated linearly. For \code{flat} extrapolation, the "outer-most"
implied volatility is used, for \code{sloped} extrapolation the slope of the two
"outer-most" implied volatilities is used.

After intra/extrapolation, the implied volatilities are transformed back into option
prices via Black&Scholes, i.e. out-of-the-money puts and calls.

Jiang & Tian (2007) define the range and increment of the strike prices in \mjseqn{SD}
units. This \mjseqn{SD} is the Black & Scholes implied volatility of the at-the-money
option. As an example, the increment recommended by Jiang & Tian (2007) is \mjseqn{0.35}
units: \mjseqn{SD \cdot \sqrt{maturity} \cdot price \cdot 0.35}. The range is calculated
accordingly as \mjseqn{SD \cdot \sqrt{maturity} \cdot price \cdot}\code{ tail_length}.
}
\examples{

library(R.MFIV)

## LOAD EXAMPLE OPTION_QUOTES
nest <- option_dataset$option_quotes[[1]]

## EXTRAPOLATE DATA
JandT_2007_smoothing_method(option_quotes = nest,
                            maturity = 0.06644802,
                            K_0 = 147,
                            R = 0.008769736,
                            F_0 = 147.5697,
                            price = 147.39)
}
\references{
\doi{10.1093/rfs/hhi027} Jiang & Tian (2005) - The Model-Free Implied
 Volatility and Its Information Content

\doi{10.3905/jod.2007.681813} Jiang & Tian (2007) - Extracting Model-Free
Volatility  from Option Prices: An Examination of the VIX Index
}
