% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{R.MFIV}
\alias{R.MFIV-package}
\alias{R.MFIV}
\title{R.MFIV: A package for handling CBOE option-data and calculation the CBOE VIX}
\description{
See the following vignettes for an explanation of how this package works:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-g-h/R.MFIV}
  \item Report bugs at \url{https://github.com/m-g-h/R.MFIV/issues}
}

}
\author{
\strong{Maintainer}: Martin Georg Haas \email{haasmartin@live.de} (\href{https://orcid.org/0000-0002-0977-0939}{ORCID})

}
\keyword{internal}
